google-sheets-agent/requirements.txt
gspread==5.12.4
oauth2client==4.1.3
pandas==2.0.3
python-dotenv==1.0.0